Add checkbox values to title/subtitle of matrix fields
Allow for the value of a checked checkbox to be used as the internal title or subtitle. Use case: A matrix where each item has a checkbox that marks it as "active." If setup correctly, the user could easily see which items were "active" by having the checkbox be the subtitle of each item.