Update router.php
PHP Warning: Undefined array key "If-Modified-Since" in D:\HTTPS-SERVER-D-VHOSTS\testi\BigTree-CMS\core\router.php on line 206

This has nothing to do with PHP versio, but with Apache configuration. My server does not send "Last-Modified" header, so there will not be a "If-Modified-Since" header, so that array key does not exist.